+++ src/base/pages/MainLayout.vue
@@ -22,6 +22,7 @@
     <div
       v-else
       class="background mainLayoutDesktop"
+      :class="isBikeKitchen ? 'bikekitchen' : ''"
     >
       <QLayout :view="layoutView">
         <QLayoutHeader reveal>
@@ -69,6 +70,7 @@
               {{ $t("GLOBAL.ABOUT_KARROT") }}
             </QItemMain>
           </QItem>
+          <CommunityFeed />
         </QLayoutDrawer>
 
         <!-- desktop sidenav -->
@@ -93,6 +95,7 @@
               {{ $t("GLOBAL.ABOUT_KARROT") }}
             </QItemMain>
           </QItem>
+          <CommunityFeed />
         </QLayoutDrawer>
 
         <QPageContainer>
@@ -160,6 +163,8 @@ import RouteError from '@/base/components/RouteError'
 import UnsupportedBrowserWarning from '@/base/components/UnsupportedBrowserWarning'
 import DetailSidebar from '@/messages/components/DetailSidebar'
 import KarrotLogo from '@/logo/components/KarrotLogo'
+import CommunityFeed from '@/communityFeed/components/CommunityFeed'
+
 import { mapGetters, mapActions } from 'vuex'
 import {
   dom,
@@ -207,6 +212,7 @@ export default {
     Banners,
     RouteError,
     UnsupportedBrowserWarning,
+    CommunityFeed,
   },
   data () {
     return {
@@ -225,6 +231,7 @@ export default {
       messagesAllUnreadMuted: 'latestMessages/allUnreadMuted',
       notificationsUnseenCount: 'notifications/unseenCount',
       currentGroupId: 'currentGroup/id',
+      isBikeKitchen: 'currentGroup/isBikeKitchen',
     }),
     layoutView () {
       if (this.$q.platform.is.mobile) {
@@ -311,6 +318,8 @@ body.desktop .mainContent-page:not(.fullpage)
   background-image url('../assets/repeating_grey.png')
   background-size: 600px
   background-attachment:fixed
+  &.bikekitchen
+    background-image url('../assets/bikekitchen_background.jpg')
 .k-highlight-dot
   position absolute
   right -4px