+++ src/feedback/components/FeedbackList.vue
@@ -7,7 +7,7 @@
     <KSpinner v-show="isPending || (feedbackPossibleStatus && feedbackPossibleStatus.pending)" />
     <KNotice v-if="empty">
       <template slot="icon">
-        <i class="fas fa-balance-scale" />
+        <i :class="$icon('feedback')" />
       </template>
       {{ $t('FEEDBACKLIST.NONE') }}
       <template slot="desc">