+++ src/sidenav/components/SidenavPlacesUI.vue
@@ -1,10 +1,10 @@
 <template>
   <SidenavBox>
     <template slot="icon">
-      <QIcon name="fas fa-fw fa-shopping-cart" />
+      <QIcon name="fas fa-fw fa-dot-circle" />
     </template>
     <template slot="name">
-      {{ $t('GROUP.STORES') }}
+      {{ $t('GROUP.PLACES') }}
     </template>
     <div
       slot="tools"