+++ src/sidenav/components/SidenavTitle.vue
@@ -7,9 +7,6 @@
     >
       {{ currentGroup && currentGroup.name }}
     </QToolbarTitle>
-    <CommunityFeed
-      @click.native="$emit('click')"
-    />
     <LatestMessageButton
       v-if="isLoggedIn"
       @click="$emit('click')"
@@ -18,7 +15,6 @@
       v-if="isLoggedIn"
       @click="$emit('click')"
     />
-    <LocaleSelect />
     <QBtn
       flat
       dense
@@ -34,8 +30,6 @@
 </template>
 
 <script>
-import LocaleSelect from '@/utils/components/LocaleSelect'
-import CommunityFeed from '@/communityFeed/components/CommunityFeed'
 import LatestMessageButton from '@/messages/components/LatestMessageButton'
 import NotificationButton from '@/notifications/components/NotificationButton'
 
@@ -56,8 +50,6 @@ export default {
     QBtn,
     QIcon,
     QTooltip,
-    LocaleSelect,
-    CommunityFeed,
     LatestMessageButton,
     NotificationButton,
   },